Alright, this was a big one. First (and least important), I fixed the issue of the snapshots bugging out when a modal was above them. This is more efficient and also keeps the page from wigging out.
The big change here is that I was sick of the scripts.js file being so long. I wanted structure. So I created a printerModule class that constructs printer modules and allows for a bunch of functionality to be encapsulated. I thought this was a great plan! And it probably is. It keeps a bunch of id passing from occurring and makes it so when you create a printer, a printer module is automatically created as well. This is great! But I think we might run into problems with naming conventions here. Technically, the printerModule is just an id with a bunch of functions. The actual printer_module DOM element has nothing to do with it except that it has the same id number (which is another problem, since currently printer ids and printer module ids are the same, but they eventually won’t be once the database is assigning them). I might need to clarify this. I might also make the actual element a child of this object, but printerModule.printerModuleDOM seems a bit weird… We will see. But we are also leaving on the break soon and I wanted to get this all figured out before we left so I don’t forget. Hopefully I have time.